feat(api-key-management-starter): align template with gallery requirements
