+++ templates/api-key-management-starter/src/services/sessionExpiredHandler.ts
@@ -1,0 +1,9 @@
+let globalSessionExpiredHandler: (() => void) | null = null;
+
+export function setGlobalSessionExpiredHandler(handler: () => void) {
+  globalSessionExpiredHandler = handler;
+}
+
+export function getGlobalSessionExpiredHandler() {
+  return globalSessionExpiredHandler;
+}
